Update docs and client structure
