+++ spec/battle_net_client/game_clients/world_of_warcraft_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-RSpec.describe BattleNetClient::Client, :vcr do
+RSpec.describe BattleNetClient::GameClients::WorldOfWarcraft, :vcr do
   subject(:client) { described_class.new }
 
   describe '#achievements' do
@@ -10,7 +10,7 @@ RSpec.describe BattleNetClient::Client, :vcr do
     let(:access_token) { 'USuiosqMobTyIawrmkMoWE47n3yMm2Kvr7' }
 
     context 'when the client is configured' do
-      before { client.region = 'eu' }
+      before { client.config.region = 'eu' }
 
       it 'returns a list of achievements' do
         expect(client.achievements(access_token:)).not_to be_empty
@@ -18,7 +18,7 @@ RSpec.describe BattleNetClient::Client, :vcr do
     end
 
     context 'when an incorrect access token is used' do
-      before { client.region = 'eu' }
+      before { client.config.region = 'eu' }
 
       it 'returns an empty array' do
         expect(client.achievements(access_token: nil)).to(be_empty)
