+++ spec/cassettes/BattleNetClient_GameClients_WorldOfWarcraft/_achievements/when_the_client_is_configured/returns_a_list_of_achievements.yml
@@ -17,7 +17,7 @@ http_interactions:
       message: OK
     headers:
       date:
-      - Mon, 31 Mar 2025 01:28:45 GMT
+      - Mon, 31 Mar 2025 03:15:47 GMT
       content-type:
       - application/json;charset=UTF-8
       transfer-encoding:
@@ -27,7 +27,7 @@ http_interactions:
       battlenet-schema-revision:
       - '1'
       last-modified:
-      - Wed, 26 Mar 2025 02:28:31 GMT
+      - Mon, 31 Mar 2025 02:28:29 GMT
       battlenet-namespace:
       - static-11.1.0_59095-eu
       vary:
@@ -41,13 +41,13 @@ http_interactions:
       content-encoding:
       - gzip
       x-trace-traceid:
-      - 3be1341c-2717-3b80-a9db-ce624de88c8d
+      - a16fea0b-cb53-3f8f-b99d-3542ed7ce3aa
       traceparent:
-      - 00-1bd41ed11f32d81160aa4a80d17d1fae-e13b5d4ebd060c55-01
+      - 00-77620f9e4423e1a7a765dbaf03f96509-146e741eb0e85c98-01
       x-trace-spanid:
-      - 1831a49c-9176-3e72-a69b-4e6eb01330de
+      - 1831a4a1-e60b-5422-0250-6a25fcd4d276
       x-trace-parentspanid:
-      - 1831a49c-9176-3e42-a69b-4e6eb01330de
+      - 1831a4a1-e60b-53f2-0250-6a25fcd4d276
       x-frame-options:
       - SAMEORIGIN
       x-content-type-options:
@@ -7318,5 +7318,5 @@ http_interactions:
         Biggest Fan","id":41694},{"key":{"href":"https://eu.api.blizzard.com/data/wow/achievement/41695?namespace=static-11.1.0_59095-eu"},"name":"Torq''s
         Biggest Fan","id":41695},{"key":{"href":"https://eu.api.blizzard.com/data/wow/achievement/41714?namespace=static-11.1.0_59095-eu"},"name":"From
         Trash to Treasure","id":41714}]}'
-  recorded_at: Mon, 31 Mar 2025 01:28:46 GMT
+  recorded_at: Mon, 31 Mar 2025 03:15:48 GMT
 recorded_with: VCR 6.3.1
